Add crumb-packed implementation of packedseq.
